+++ StayWell/src/main/java/com/staywell/service/ReservationService.java
@@ -9,14 +9,14 @@ import com.staywell.model.Reservation;
 
 public interface ReservationService {
 
-	public Reservation createReservation(Integer roomId, ReservationDTO reservationDTO, String paymentType, String txnId) throws ReservationException, RoomException;
+	public Reservation createReservation(Long roomId, ReservationDTO reservationDTO, String paymentType, String txnId) throws ReservationException, RoomException;
 	
-	public String cancelReservation(Integer reservationId) throws ReservationException;
+	public String cancelReservation(Long reservationId) throws ReservationException;
 	
 	public List<Reservation> getAllReservationsOfHotel() throws ReservationException;
 	
 	public List<Reservation> getAllReservationsOfCustomer() throws ReservationException;
 	
-	public Reservation getReservationById(Integer ReservationId) throws ReservationException;
+	public Reservation getReservationById(Long ReservationId) throws ReservationException;
 
 }
